Moving test config to test folder
